Latest content on home page
Converted the test winner of #446 to our source code. Meaning the search bar is replaced with buttons and the latest content section is added. In order for it to work I also added a new button (the blue one) to the css and updated the publishing data of the most recent two pages