Removed changes that did not work
Removed the changes that did not work. Namely remove image on homepage for mobile, put linkedin script back to async instead of defer and removed the style for alerts in the header. Lastly put the optimize script in notations, so if a/b tests want to be done put this back on