Upload first version of kableExtra building block
The task description asks for an example to further customize modelsummary output with kableExtra before outputting the table to Latex. So far, this is indeed a right example to show the functions kableExtra provides I think. However, the table is outputted to HTML. I did not manage yet to get it produce Latex code.